fix WIN build fail
